requirements.txt

channels
channels-redis
daphne
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
python-decouple
redis